+++ src/components/expense-types/forms/sections/Details.vue
